update: Split Sound_Class into Speaker_Class and Mic_Class.
