+++ src/M5Unified.cpp
@@ -38,6 +38,62 @@ void __attribute((weak)) adc_power_acquire(void)
 
 namespace m5
 {
+  bool M5Unified::_speaker_enabled_cb(void* args, bool enabled)
+  {
+    auto self = (M5Unified*)args;
+
+    switch (self->getBoard())
+    {
+#if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
+    case board_t::board_M5StackCore2:
+    case board_t::board_M5Tough:
+      self->Power.Axp192.setGPIO2(enabled);
+      break;
+
+    case board_t::board_M5StickC:
+    case board_t::board_M5StickCPlus:
+    case board_t::board_M5StackCoreInk:
+      /// for SPK HAT
+      if ((self->_cfg.external_spk_detail.enabled) && !self->_cfg.external_spk_detail.omit_spk_hat)
+      {
+        gpio_num_t pin_en = self->_board == board_t::board_M5StackCoreInk ? GPIO_NUM_25 : GPIO_NUM_0;
+        if (enabled)
+        {
+          m5gfx::pinMode(pin_en, m5gfx::pin_mode_t::output);
+          m5gfx::gpio_hi(pin_en);
+        }
+        else
+        { m5gfx::gpio_lo(pin_en); }
+      }
+      break;
+
+#endif
+    default:
+      break;
+    }
+    return true;
+  }
+
+  bool M5Unified::_microphone_enabled_cb(void* args, bool enabled)
+  {
+    auto self = (M5Unified*)args;
+
+    switch (self->getBoard())
+    {
+#if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
+    case board_t::board_M5StickC:
+    case board_t::board_M5StickCPlus:
+      self->Power.Axp192.setLDO0(enabled ? 2800 : 0);
+      break;
+
+#endif
+    default:
+      break;
+    }
+    return true;
+  }
+
+/*
   bool M5Unified::_sound_set_mode_cb(void* args, m5::sound_mode_t mode)
   {
     auto self = (M5Unified*)args;
@@ -78,7 +134,7 @@ namespace m5
     }
     return true;
   }
-
+*/
 
 
 #if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
@@ -273,152 +329,159 @@ namespace m5
 
 #endif
 
-    if (_cfg.internal_spk || _cfg.external_spk_detail.enabled || _cfg.internal_mic)
+    if (_cfg.internal_mic)
     {
-      auto sound_cfg = Sound.config();
+      auto mic_cfg = Mic.config();
 
-      if (_cfg.internal_mic)
+      // mic_cfg.magnification = 4;
+      mic_cfg.over_sampling = 2;
+      switch (_board)
       {
-        sound_cfg.mic_gain = 10;
-        sound_cfg.mic_over_sampling = 2;
-        switch (_board)
-        {
 #if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
-        case board_t::board_M5Stack:
-          if (_cfg.internal_mic)
-          {
-            sound_cfg.pin_data_in = 34;  // M5GO bottom MIC
-            sound_cfg.mic_adc = true;
-            sound_cfg.mic_offset = 192;
-            sound_cfg.mic_over_sampling = 4;
-          }
-          break;
-
-        case board_t::board_M5StickC:
-        case board_t::board_M5StickCPlus:
-        case board_t::board_M5Tough:
-        case board_t::board_M5StackCore2:
-          if (_cfg.internal_mic)
-          { /// builtin PDM mic
-            sound_cfg.pin_data_in = 34;
-            sound_cfg.pin_lrck = 0;
-          }
-          break;
-
-        case board_t::board_M5AtomU:
-          { /// ATOM U builtin PDM mic
-            sound_cfg.pin_lrck = 5;
-            sound_cfg.pin_data_in = 19;
-            sound_cfg.mic_offset = - 768;
-          }
-          break;
-
-        case board_t::board_M5Atom:
-          { /// ATOM ECHO builtin PDM mic
-            sound_cfg.pin_lrck = 33;
-            sound_cfg.pin_data_in = 23;
-          }
-          break;
-#endif
-        default:
-          break;
+      case board_t::board_M5Stack:
+        if (_cfg.internal_mic)
+        {
+          mic_cfg.pin_data_in = 34;  // M5GO bottom MIC
+          mic_cfg.use_adc = true;    // use ADC analog input
+          mic_cfg.input_offset = 192;
+          mic_cfg.over_sampling = 4;
         }
-      }
+        break;
 
-      if (_cfg.internal_spk || _cfg.external_spk_detail.enabled)
+      case board_t::board_M5StickC:
+      case board_t::board_M5StickCPlus:
+      case board_t::board_M5Tough:
+      case board_t::board_M5StackCore2:
+        if (_cfg.internal_mic)
+        { /// builtin PDM mic
+          mic_cfg.pin_data_in = 34;
+          mic_cfg.pin_ws = 0;
+        }
+        break;
+
+      case board_t::board_M5AtomU:
+        { /// ATOM U builtin PDM mic
+          mic_cfg.pin_data_in = 19;
+          mic_cfg.pin_ws = 5;
+          mic_cfg.input_offset = - 768;
+        }
+        break;
+
+      case board_t::board_M5Atom:
+        { /// ATOM ECHO builtin PDM mic
+          mic_cfg.pin_data_in = 23;
+          mic_cfg.pin_ws = 33;
+        }
+        break;
+#endif
+      default:
+        break;
+      }
+      if (mic_cfg.pin_data_in >= 0)
       {
-        // set default speaker gain.
-        sound_cfg.spk_gain = 32;
-        switch (_board)
-        {
-#if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
-        case board_t::board_M5Stack:
-          if (_cfg.internal_spk)
-          {
-            m5gfx::gpio_lo(GPIO_NUM_25);
-            m5gfx::pinMode(GPIO_NUM_25, m5gfx::pin_mode_t::output);
-            sound_cfg.spk_dac = true;
-            sound_cfg.pin_data_out = 25;
-          }
-          break;
-
-        case board_t::board_M5StackCoreInk:
-        case board_t::board_M5StickCPlus:
-          if (_cfg.internal_spk)
-          {
-            sound_cfg.spk_buzzer = true;
-            sound_cfg.pin_data_out = 2;
-            // Since these buzzers are very small, we have set the gain to the maximum value here.
-            sound_cfg.spk_gain = 255;
-          }
-          NON_BREAK;
-        case board_t::board_M5StickC:
-          if (_cfg.external_spk_detail.enabled && !_cfg.external_spk_detail.omit_spk_hat)
-          { /// for SPK HAT
-            gpio_num_t pin_en = _board == board_t::board_M5StackCoreInk ? GPIO_NUM_25 : GPIO_NUM_0;
-            m5gfx::gpio_lo(pin_en);
-            m5gfx::pinMode(pin_en, m5gfx::pin_mode_t::output);
-            m5gfx::gpio_lo(GPIO_NUM_26);
-            m5gfx::pinMode(GPIO_NUM_26, m5gfx::pin_mode_t::output);
-            sound_cfg.pin_data_out = 26;
-            sound_cfg.spk_dac = true;
-            sound_cfg.spk_buzzer = false;
-            sound_cfg.spk_gain = 128;
-          }
-          break;
-
-        case board_t::board_M5Tough:
-          // The gain is set higher than Core2 here because the waterproof case reduces the sound.;
-          sound_cfg.spk_gain = 64;
-          NON_BREAK;
-        case board_t::board_M5StackCore2:
-          if (_cfg.internal_spk)
-          {
-            sound_cfg.pin_bck = 12;
-            sound_cfg.pin_lrck = 0;
-            sound_cfg.pin_data_out = 2;
-          }
-          break;
-
-        case board_t::board_M5Atom:
-          if (_cfg.internal_spk)
-          { // for ATOM ECHO
-            sound_cfg.pin_bck = 19;
-            sound_cfg.pin_lrck = 33;
-            sound_cfg.pin_data_out = 22;
-            sound_cfg.spk_gain = 16;
-          }
-          NON_BREAK;
-        case board_t::board_M5AtomPsram:
-          if (_cfg.external_spk_detail.enabled && !_cfg.external_spk_detail.omit_atomic_spk)
-          { // for ATOMIC SPK
-            // 19,23,33 pulldown read check ( all high = ATOMIC_SPK ? )
-            gpio_num_t pin = (_board == board_t::board_M5AtomPsram) ? GPIO_NUM_5 : GPIO_NUM_23;
-            m5gfx::pinMode(GPIO_NUM_19, m5gfx::pin_mode_t::input_pulldown);
-            m5gfx::pinMode(GPIO_NUM_33, m5gfx::pin_mode_t::input_pulldown);
-            m5gfx::pinMode(pin        , m5gfx::pin_mode_t::input_pulldown);
-            if (m5gfx::gpio_in(GPIO_NUM_19)
-             && m5gfx::gpio_in(GPIO_NUM_33)
-             && m5gfx::gpio_in(pin        ))
-            {
-              _cfg.internal_imu = false; /// avoid conflict with i2c
-              _cfg.internal_rtc = false; /// avoid conflict with i2c
-              sound_cfg.pin_bck = 22;
-              sound_cfg.pin_lrck = 21;
-              sound_cfg.pin_data_out = 25;
-              sound_cfg.pin_data_in = -1;   // disable mic for ECHO
-              sound_cfg.spk_gain = 32;
-            }
-          }
-          break;
-#endif
-        default:
-          break;
-        }
+        Mic.setCallback(this, _microphone_enabled_cb);
+        Mic.config(mic_cfg);
       }
+    }
 
-      Sound.setCallback(this, _sound_set_mode_cb);
-      Sound.config(sound_cfg);
+    if (_cfg.internal_spk || _cfg.external_spk_detail.enabled)
+    {
+      auto spk_cfg = Speaker.config();
+      // set default speaker gain.
+      spk_cfg.magnification = 16;
+      switch (_board)
+      {
+#if !defined (CONFIG_IDF_TARGET) || defined (CONFIG_IDF_TARGET_ESP32)
+      case board_t::board_M5Stack:
+        if (_cfg.internal_spk)
+        {
+          m5gfx::gpio_lo(GPIO_NUM_25);
+          m5gfx::pinMode(GPIO_NUM_25, m5gfx::pin_mode_t::output);
+          spk_cfg.use_dac = true;
+          spk_cfg.pin_data_out = 25;
+          spk_cfg.magnification = 8;
+        }
+        break;
+
+      case board_t::board_M5StackCoreInk:
+      case board_t::board_M5StickCPlus:
+        if (_cfg.internal_spk)
+        {
+          spk_cfg.buzzer = true;
+          spk_cfg.pin_data_out = 2;
+          spk_cfg.magnification = 64;
+        }
+        NON_BREAK;
+      case board_t::board_M5StickC:
+        if (_cfg.external_spk_detail.enabled && !_cfg.external_spk_detail.omit_spk_hat)
+        { /// for SPK HAT
+          gpio_num_t pin_en = _board == board_t::board_M5StackCoreInk ? GPIO_NUM_25 : GPIO_NUM_0;
+          m5gfx::gpio_lo(pin_en);
+          m5gfx::pinMode(pin_en, m5gfx::pin_mode_t::output);
+          m5gfx::gpio_lo(GPIO_NUM_26);
+          m5gfx::pinMode(GPIO_NUM_26, m5gfx::pin_mode_t::output);
+          spk_cfg.pin_data_out = 26;
+          spk_cfg.use_dac = true;
+          spk_cfg.buzzer = false;
+          spk_cfg.magnification = 32;
+        }
+        break;
+
+      case board_t::board_M5Tough:
+        // The gain is set higher than Core2 here because the waterproof case reduces the sound.;
+        spk_cfg.magnification = 32;
+        NON_BREAK;
+      case board_t::board_M5StackCore2:
+        if (_cfg.internal_spk)
+        {
+          spk_cfg.pin_bck = 12;
+          spk_cfg.pin_ws = 0;
+          spk_cfg.pin_data_out = 2;
+        }
+        break;
+
+      case board_t::board_M5Atom:
+        if (_cfg.internal_spk)
+        { // for ATOM ECHO
+          spk_cfg.pin_bck = 19;
+          spk_cfg.pin_ws = 33;
+          spk_cfg.pin_data_out = 22;
+          spk_cfg.magnification = 12;
+        }
+        NON_BREAK;
+      case board_t::board_M5AtomPsram:
+        if (_cfg.external_spk_detail.enabled && !_cfg.external_spk_detail.omit_atomic_spk)
+        { // for ATOMIC SPK
+          // 19,23,33 pulldown read check ( all high = ATOMIC_SPK ? )
+          gpio_num_t pin = (_board == board_t::board_M5AtomPsram) ? GPIO_NUM_5 : GPIO_NUM_23;
+          m5gfx::pinMode(GPIO_NUM_19, m5gfx::pin_mode_t::input_pulldown);
+          m5gfx::pinMode(GPIO_NUM_33, m5gfx::pin_mode_t::input_pulldown);
+          m5gfx::pinMode(pin        , m5gfx::pin_mode_t::input_pulldown);
+          if (m5gfx::gpio_in(GPIO_NUM_19)
+            && m5gfx::gpio_in(GPIO_NUM_33)
+            && m5gfx::gpio_in(pin        ))
+          {
+            _cfg.internal_imu = false; /// avoid conflict with i2c
+            _cfg.internal_rtc = false; /// avoid conflict with i2c
+            spk_cfg.pin_bck = 22;
+            spk_cfg.pin_ws = 21;
+            spk_cfg.pin_data_out = 25;
+            spk_cfg.magnification = 16;
+            auto mic = Mic.config();
+            mic.pin_data_in = -1;   // disable mic for ECHO
+            Mic.config(mic);
+          }
+        }
+        break;
+#endif
+      default:
+        break;
+      }
+      if (spk_cfg.pin_data_out >= 0)
+      {
+        Speaker.setCallback(this, _speaker_enabled_cb);
+        Speaker.config(spk_cfg);
+      }
     }
 
     switch (_board) /// setup Hardware Buttons
