+++ src/utility/Speaker_Class.hpp
@@ -1,8 +1,8 @@
 // Copyright (c) M5Stack. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-#ifndef __M5_Sound_Class_H__
-#define __M5_Sound_Class_H__
+#ifndef __M5_Speaker_Class_H__
+#define __M5_Speaker_Class_H__
 
 #include <freertos/FreeRTOS.h>
 #include <freertos/task.h>
@@ -12,61 +12,43 @@ namespace m5
 {
   class M5Unified;
 
-  enum sound_mode_t
-  {
-    sound_off,
-    sound_output,
-    sound_input,
-  };
-
-  struct sound_config_t
+  struct speaker_config_t
   {
     /// i2s_data_out (for spk)
-    int pin_data_out = -1;
-
-    /// i2s_data_in (for mic)
-    int pin_data_in = -1;
+    int pin_data_out = I2S_PIN_NO_CHANGE;
 
     /// i2s_bck
-    int pin_bck = -1;
+    int pin_bck = I2S_PIN_NO_CHANGE;
 
-    /// i2s_lrck
-    int pin_lrck = -1;
+    /// i2s_ws (lrck)
+    int pin_ws = I2S_PIN_NO_CHANGE;
 
-    int spk_sample_rate = 62500;
+    /// output sampling rate (Hz)
+    uint32_t sample_rate = 50000;
 
     /// use stereo output
-    bool spk_stereo = false;
+    bool stereo = false;
 
-    /// use single gpio buzzer, need set pin_data_out.
-    bool spk_buzzer = false;
+    /// use single gpio buzzer, ( need only pin_data_out )
+    bool buzzer = false;
 
-    /// use DAC speaker, need set pin_data_out ( only 25 or 26 )
-    bool spk_dac = false;
+    /// use DAC speaker, ( need only pin_data_out ) ( only GPIO_NUM_25 or GPIO_NUM_26 )
+    bool use_dac = false;
 
     /// multiplier for output value
-    uint8_t spk_gain = 32;
+    uint8_t magnification = 16;
 
-    /// use analog input mic ( need only pin_data_in )
-    bool mic_adc = false;
-
-    /// offset correction value of ADC input value
-    int mic_offset = 0;
-
-    uint8_t mic_over_sampling = 4;
-
-    int mic_sample_rate = 16000;
-
-    /// multiplier for input value
-    uint8_t mic_gain = 10;
-
+    /// background task priority
     int task_priority = 2;
+
+    /// background task pinned core
     int task_pinned_core = 0;
+
+    /// I2S port
     i2s_port_t i2s_port = i2s_port_t::I2S_NUM_0;
   };
 
-
-  class Sound_Class
+  class Speaker_Class
   {
   friend M5Unified;
 
@@ -76,11 +58,16 @@ namespace m5
 
   public:
 
-    sound_config_t config(void) const { return _cfg; }
-    void config(const sound_config_t& cfg) { _cfg = cfg; }
+    speaker_config_t config(void) const { return _cfg; }
+    void config(const speaker_config_t& cfg) { _cfg = cfg; }
 
-    bool hasSpk(void) const { return _cfg.pin_data_out >= 0; }
-    bool hasMic(void) const { return _cfg.pin_data_in  >= 0; }
+    bool begin(void);
+
+    void end(void);
+
+    bool isRunning(void) const { return _task_running; }
+
+    bool isEnabled(void) const { return _cfg.pin_data_out >= 0; }
 
     /// now in playing or not.
     /// @return false=not playing / true=playing
@@ -90,10 +77,6 @@ namespace m5
     /// @param channel virtual channel number. (0~7), (default = automatically selected)
     /// @return 0=not playing / 1=playing (There's room in the queue) / 2=playing (There's no room in the queue.)
     size_t isPlaying(uint8_t channel) const { return (channel < sound_channel_max) ? (bool)_ch_info[channel].next_wav.repeat + (bool)_ch_info[channel].current_wav.repeat : 0; }
-
-    /// now in recording or not.
-    /// @return 0=not recording / 1=recording (There's room in the queue) / 2=recording (There's no room in the queue.)
-    size_t isRecording(void) const { return _is_recording ? ((bool)_rec_info[0].length) + ((bool)_rec_info[1].length) : 0; }
 
     /// sets the output master volume of the sound.
     /// @param master_volume master volume (0~255)
@@ -118,35 +101,36 @@ namespace m5
     uint8_t getChannelVolume(uint8_t channel) const { return (channel < sound_channel_max) ? _ch_info[channel].volume : 0; }
 
     /// stop sound output.
-    void stopPlay(void);
+    void stop(void);
 
     /// stop sound output for the specified virtual channel.
     /// @param channel virtual channel number. (0~7)
-    void stopPlay(uint8_t channel);
+    void stop(uint8_t channel);
 
     /// play simple tone sound.
     /// @param frequency tone frequency (Hz)
     /// @param duration tone duration (msec)
     /// @param channel virtual channel number. (0~7), (default = automatically selected)
+    /// @param stop_current_sound true=start a new output without waiting for the current one to finish.
     /// @param wav_data Single amplitude audio data. 8bit unsigned wav.
     /// @param array_len size of wav_data.
     /// @param stereo true=data is stereo / false=data is mono.
-    bool tone(float frequency, uint32_t duration, int channel, const uint8_t* wav_data, size_t array_len, bool stereo = false);
+    bool tone(float frequency, uint32_t duration, int channel, bool stop_current_sound, const uint8_t* wav_data, size_t array_len, bool stereo = false);
 
     /// play simple tone sound.
     /// @param frequency tone frequency (Hz)
     /// @param duration tone duration (msec)
     /// @param channel virtual channel number. (0~7), (default = automatically selected)
-    bool tone(float frequency, uint32_t duration = ~0u, int channel = -1) { return tone(frequency, duration, channel, _default_tone_wav, sizeof(_default_tone_wav), false); }
+    bool tone(float frequency, uint32_t duration = ~0u, int channel = -1, bool stop_current_sound = true) { return tone(frequency, duration, channel, stop_current_sound, _default_tone_wav, sizeof(_default_tone_wav), false); }
 
     /// play raw sound wave data. (for signed 8bit wav data)
     /// @param wav_data wave data.
     /// @param array_len Number of data array elements.
     /// @param sample_rate the sampling rate (Hz) (default = 44100)
     /// @param stereo true=data is stereo / false=data is monaural.
-    /// @param repeat : Number of times played repeatedly. (default = 1)
-    /// @param channel : virtual channel number (If omitted, use an available channel.)
-    /// @param stop_current_sound : true=Start a new output without waiting for the current one to finish.
+    /// @param repeat number of times played repeatedly. (default = 1)
+    /// @param channel virtual channel number (If omitted, use an available channel.)
+    /// @param stop_current_sound true=start a new output without waiting for the current one to finish.
     bool playRAW(const int8_t* wav_data, size_t array_len, uint32_t sample_rate = 44100, bool stereo = false, uint32_t repeat = 1, int channel = -1, bool stop_current_sound = false)
     {
       return _play_raw(static_cast<const void* >(wav_data), array_len, false, true, sample_rate, stereo, repeat, channel, stop_current_sound, false);
@@ -157,9 +141,9 @@ namespace m5
     /// @param array_len Number of data array elements.
     /// @param sample_rate the sampling rate (Hz) (default = 44100)
     /// @param stereo true=data is stereo / false=data is monaural.
-    /// @param repeat : Number of times played repeatedly. (default = 1)
-    /// @param channel : virtual channel number (If omitted, use an available channel.)
-    /// @param stop_current_sound : true=Start a new output without waiting for the current one to finish.
+    /// @param repeat number of times played repeatedly. (default = 1)
+    /// @param channel virtual channel number (If omitted, use an available channel.)
+    /// @param stop_current_sound true=start a new output without waiting for the current one to finish.
     bool playRAW(const uint8_t* wav_data, size_t array_len, uint32_t sample_rate = 44100, bool stereo = false, uint32_t repeat = 1, int channel = -1, bool stop_current_sound = false)
     {
       return _play_raw(static_cast<const void* >(wav_data), array_len, false, false, sample_rate, stereo, repeat, channel, stop_current_sound, false);
@@ -170,62 +154,22 @@ namespace m5
     /// @param array_len Number of data array elements.
     /// @param sample_rate the sampling rate (Hz) (default = 44100)
     /// @param stereo true=data is stereo / false=data is monaural.
-    /// @param repeat : Number of times played repeatedly. (default = 1)
-    /// @param channel : virtual channel number (If omitted, use an available channel.)
-    /// @param stop_current_sound : true=Start a new output without waiting for the current one to finish.
+    /// @param repeat number of times played repeatedly. (default = 1)
+    /// @param channel virtual channel number (If omitted, use an available channel.)
+    /// @param stop_current_sound true=start a new output without waiting for the current one to finish.
     bool playRAW(const int16_t* wav_data, size_t array_len, uint32_t sample_rate = 44100, bool stereo = false, uint32_t repeat = 1, int channel = -1, bool stop_current_sound = false)
     {
       return _play_raw(static_cast<const void* >(wav_data), array_len, true, true, sample_rate, stereo, repeat, channel, stop_current_sound, false);
     }
 
-    /// set recording sampling rate.
-    /// @param sample_rate the sampling rate (Hz)
-    void setRecordSampleRate(uint32_t sample_rate) { _cfg.mic_sample_rate = sample_rate; }
-
-    /// record raw sound wave data.
-    /// @param rec_data Recording destination array.
-    /// @param array_len Number of data array elements.
-    /// @param sample_rate the sampling rate (Hz)
-    bool record(uint8_t* rec_data, size_t array_len, uint32_t sample_rate)
-    {
-      return _rec_raw(rec_data, array_len, false, sample_rate);
-    }
-
-    /// record raw sound wave data.
-    /// @param rec_data Recording destination array.
-    /// @param array_len Number of data array elements.
-    /// @param sample_rate the sampling rate (Hz)
-    bool record(int16_t* rec_data, size_t array_len, uint32_t sample_rate)
-    {
-      return _rec_raw(rec_data, array_len,  true, sample_rate);
-    }
-
-    /// record raw sound wave data.
-    /// @param rec_data Recording destination array.
-    /// @param array_len Number of data array elements.
-    bool record(uint8_t* rec_data, size_t array_len)
-    {
-      return _rec_raw(rec_data, array_len, false, _cfg.mic_sample_rate);
-    }
-
-    /// record raw sound wave data.
-    /// @param rec_data Recording destination array.
-    /// @param array_len Number of data array elements.
-    bool record(int16_t* rec_data, size_t array_len)
-    {
-      return _rec_raw(rec_data, array_len,  true, _cfg.mic_sample_rate);
-    }
-
-    void setCallback(void* args, bool(*func)(void*, sound_mode_t)) { _cb_set_mode = func; _cb_set_mode_args = args; }
-
-    bool setMode(sound_mode_t mode);
-
   protected:
+
+    void setCallback(void* args, bool(*func)(void*, bool)) { _cb_set_enabled = func; _cb_set_enabled_args = args; }
 
     struct wav_info_t
     {
       size_t length = 0;
-      int repeat = 0;   /// -1 mean infinity repeat
+      uint32_t repeat = 0;   /// -1 mean infinity repeat
       uint32_t sample_rate = 0;
       const void* data = nullptr;
       union
@@ -250,43 +194,26 @@ namespace m5
       wav_info_t current_wav;
       wav_info_t next_wav;
       size_t index = 0;
-      int32_t diff = 0;
-      uint8_t volume = 32; // channel volume (not master volume)
+      size_t diff = 0;
+      uint8_t volume = 64; // channel volume (not master volume)
     };
 
     channel_info_t _ch_info[sound_channel_max];
 
-    struct recording_info_t
-    {
-      void* data = nullptr;
-      size_t length = 0;
-      bool is_16bit = false;
-    };
-
-    recording_info_t _rec_info[2];
-
     static void output_task(void* args);
-    static void input_task(void* args);
 
-    int _calc_rec_rate(void) const;
-    esp_err_t _setup_i2s(bool mic);
+    esp_err_t _setup_i2s(void);
     bool _play_raw(const void* wav, size_t array_len, bool flg_16bit, bool flg_signed, uint32_t sample_rate, bool flg_stereo, uint32_t repeat_count, int channel, bool stop_current_sound, bool no_clear_index);
-    bool _rec_raw(void* recdata, size_t array_len, bool flg_16bit, uint32_t sample_rate);
 
-    sound_mode_t _sound_mode = sound_off;
-    i2s_dac_mode_t _dac_mode = i2s_dac_mode_t::I2S_DAC_CHANNEL_DISABLE;
+    speaker_config_t _cfg;
+    uint8_t _master_volume = 64;
 
-    sound_config_t _cfg;
-    uint8_t _master_volume = 32;
-
-    bool (*_cb_set_mode)(void* args, sound_mode_t mode) = nullptr;
-    void* _cb_set_mode_args = nullptr;
+    bool (*_cb_set_enabled)(void* args, bool enabled) = nullptr;
+    void* _cb_set_enabled_args = nullptr;
 
     TaskHandle_t _sound_task_handle = nullptr;
-    volatile uint8_t _play_channel_bits = 0;
     volatile bool _task_running = false;
-    volatile bool _is_recording = false;
-    uint32_t _rec_sample_rate = 0;
+    volatile uint8_t _play_channel_bits = 0;
   };
 }
